backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
pydantic==2.10.0
python-dotenv==1.0.1
requests==2.31.0
